django-test/requirements.txt
django
djangorestframework
psycopg2-binary
gunicron
django_heroku